Refactor #358 - migrate logo to designSystem
