+++ core/designsystem/src/main/kotlin/com/nlab/reminder/core/designsystem/compose/icon/PlaneatIcons.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2023 The N's lab Open Source Project
+ * Copyright (C) 2024 The N's lab Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,9 +14,17 @@
  * limitations under the License.
  */
 
-package com.nlab.reminder.core.android.designsystem.icon
+package com.nlab.reminder.core.designsystem.compose.icon
+
+import com.nlab.reminder.core.designsystem.compose.resource.imagevector.IcHomeCategoryAll
+import com.nlab.reminder.core.designsystem.compose.resource.imagevector.IcHomeCategoryTimetable
+import com.nlab.reminder.core.designsystem.compose.resource.imagevector.IcHomeCategoryToday
 
 /**
- * @author thalys
+ * @author Doohyun
  */
-object PlanetIcons
+object PlaneatIcons {
+    val todayLogo = IcHomeCategoryToday
+    val timetableLogo = IcHomeCategoryTimetable
+    val allLogo = IcHomeCategoryAll
+}
